Minor improvement to fzf install path logic and startup file copy

+++ to_scripts/install-vpp.sh
@@ -73,15 +73,20 @@ if [ ! -d vpp-scripts ]; then
     echo "ERROR: Failed to clone vpp-scripts repo."
     exit 1
   fi
+else
+  echo "Bypassing vpp-scripts because it already exists."
+fi
+cd ~
+if [ ! -f startup.conf ]; then
   cp vpp-scripts/startup.conf.example startup.conf
-  cp vpp-scripts/startup_cmds.vpp startup_cmds.vpp
   # Replace USER placeholder with actual username
   sed -i "s/\bUSER\b/$USER/g" startup.conf || {
     echo "ERROR: Failed to replace USER in startup.conf"
     exit 1
   }
-else
-  echo "Bypassing vpp-scripts because it already exists."
+fi
+if [ ! -f startup_cmds.vpp ]; then
+  cp vpp-scripts/startup_cmds.vpp startup_cmds.vpp
 fi
 
 if [ ! -f ~/vppctl ]; then
